Set cuda decoder's HostMatrix allocations to zero after creation.
When you create an instance of CudaDecoder, delete it, and then create
another instance of it, it is common that the past memory allocations
get reused. This can cause errors like seg faults in the second
CudaDecoder instance.

It is certain that there is a subtle logic error somewhere within the
code, but I can't spot it after two days of inspecting. This is an
easy solution without a meaningful performance penalty, meanwhile.